added typescript along with types, packages with their types, and a partial converstion to ts/tsx
