--- components/Food/AllFood.js
@@ -1,15 +1,0 @@
-import EachFood from "./EachFood";
-
-const AllFoods = ({ food, setFood }) => {
-  return (
-    <>
-      {food.map((item) => (
-        <div key={item._id}>
-          <EachFood food={item} setFood={setFood} />
-        </div>
-      ))}
-    </>
-  );
-};
-
-export default AllFoods;
