+++ components/Layout.tsx
@@ -1,0 +1,13 @@
+import Navigation from 'components/Navigation';
+import { ReactNode } from 'react';
+// import Footer from './Footer';
+
+export default function Layout({ children }: { children: ReactNode }) {
+  return (
+    <>
+      <Navigation />
+      <main>{children}</main>
+      {/* <Footer /> */}
+    </>
+  );
+}
