--- components/Loading.js
@@ -1,16 +1,0 @@
-import { LoadingOverlay } from "@mantine/core";
-import { Loader } from '@mantine/core';
-
-import styles from '../styles/Loader.module.css';
-
-export default function Loading(props) {
-
-    // const [isLoading, setIsLoading] = useState(true)
-
-    // return <LoadingOverlay visible={props.isLoading} />
-    return (
-        <div className={styles['container']}>
-            <Loader color="dark" loading={props.loading} />
-        </div>
-    )
-}
