+++ package.json
@@ -6,7 +6,9 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "d": "next dev",
+    "bs": "next build && next start"
   },
   "dependencies": {
     "@emotion/react": "^11.10.6",
@@ -18,6 +20,19 @@
     "next": "13.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "vercel": "^28.18.1"
+    "react-hook-form": "^7.43.7",
+    "sass": "^1.49.10",
+    "sharp": "^0.30.6",
+    "vercel": "^28.18.1",
+    "yup": "^1.0.2"
+  },
+  "devDependencies": {
+    "@types/node": "17.0.23",
+    "@types/yup": "^0.32.0",
+    "@types/react": "17.0.43",
+    "@types/react-dom": "17.0.14",
+    "eslint": "8.12.0",
+    "eslint-config-next": "12.1.3",
+    "typescript": "^5.0.2"
   }
 }
