--- pages/_app.js
@@ -1,10 +1,0 @@
-import Layout from '../components/Layout';
-import '@/styles/globals.css';
-
-export default function App({ Component, pageProps }) {
-  return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
-  );
-}
