+++ pages/_app.tsx
@@ -1,0 +1,11 @@
+import Layout from 'components/Layout';
+import { AppProps } from 'next/app';
+import '/styles/globals.css';
+
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  );
+}
