--- pages/_document.js
@@ -1,15 +1,0 @@
-import Document, { Head, Html, Main, NextScript } from 'next/document';
-
-export default class _Document extends Document {
-  render() {
-    return (
-      <Html>
-        <Head />
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
