+++ pages/_document.tsx
@@ -1,0 +1,15 @@
+import Document, { Head, Html, Main, NextScript } from 'next/document';
+
+export default class _Document extends Document {
+  render() {
+    return (
+      <Html>
+        <Head />
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    );
+  }
+}
