+++ settings/cfg.settings.ts
@@ -1,0 +1,1 @@
+export const cfg = {};
